Trim ui/math.rs: split widget out, retire static-Once CSS injection
Closes #35. Wave 3 architectural cleanup.

`ui/math.rs` was 493 lines holding three separable concerns:
1. Pure evaluator (DrawerOpsFactory, eval_expression, format_result, +
   30 unit tests covering migration regressions and format branches)
2. GTK widget builder (build_math_result, append_copy_button,
   keyboard-nav controller)
3. Dynamic CSS injection via `static Once`, out of step with the rest
   of the codebase (everything else uses assets/drawer.css +
   include_str!)

After the split:

- `ui/math.rs` (358 lines, was 493) — evaluator + tests only. Zero
  GTK deps; fully unit-testable.
- `ui/math_widget.rs` (133 lines, new) — `build_math_result` +
  `append_copy_button` + keyboard-nav controller. Calls
  `super::math::eval_expression` / `format_result`.
- `assets/drawer.css` (+34 lines) — gains `.math-result` /
  `.math-divider` / `.math-copy` / `.math-copied` rules, with values
  baked in alongside the explanatory section comment.
- `ui/constants.rs` (-8 lines) — five CSS-only constants
  (MATH_FONT_SIZE, MATH_SPACING, MATH_BORDER_RADIUS,
  MATH_BUTTON_PADDING_V, MATH_BUTTON_PADDING_H) removed; they were
  only consumed by the now-dead format!-into-CSS template. The
  remaining MATH_VBOX_SPACING is still used by the Rust-side vbox
  construction.

The `static Once` block is gone — math styles load with the rest of
the embedded stylesheet at activate time.

`format_result` is now `pub(super)` so `math_widget` can call it.
The 1 caller of `build_math_result` (well_builder.rs) updated to
`ui::math_widget::build_math_result(phrase)`.

CLAUDE.md "What lives where" tree updated.

`make lint` clean. All 82 tests pass — including the full math test
suite that stays with the evaluator.
